rpk: split plugin commands in separate files
